# api/v1/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TutorPlan Website RestFul API",
    "version": "1.0.0"
  },
  "basePath": "/",
...
  "paths": {
...
    "/api/v1/availability": {
      "get": {
        "summary": "Get all availability",
        "description": "Retrieve all availability",
        "tags": ["Availability"],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Availability"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new availability",
        "description": "Add a new availability",
        "tags": ["Availability"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "availability",
            "in": "body",
            "description": "Availability object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Availability"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Availability added successfully",
            "schema": {
              "$ref": "#/definitions/Availability"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Invalid course_id"
          }
        }
      }
    },
    "/api/v1/availability/{course_id}/booked": {
      "get": {
        "summary": "Get booked availability of a course",
        "description": "Retrieve all booked availability of a specific course",
        "tags": ["Availability"],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "description": "ID of the course",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Availability"
              }
            }
          },
          "404": {
            "description": "Course not found"
          }
        }
      }
    },
    "/api/v1/availability/{course_id}/unbooked": {
      "get": {
        "summary": "Get unbooked availability of a course",
        "description": "Retrieve all unbooked availability of a specific course",
        "tags": ["Availability"],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "description": "ID of the course",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Availability"
              }
            }
          },
          "404": {
            "description": "Course not found"
          }
        }
      }
    },
    "/api/v1/availability/{course_id}": {
      "get": {
        "summary": "Get all availabilities of a course",
        "description": "Retrieve all availabilities of a course.",
        "tags": ["Availability"],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "description": "ID of the course",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Availability"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Course not found"
          }
        }
      },      
      "delete": {
        "summary": "Delete availability or availabilities of a course",
        "description": "Delete the availability  or availabilities of a course if it has not been booked",
        "tags": ["Availability"],
        "consumes": ["application/json"],

        "parameters": [
          {
            "name": "availability ids",
            "in": "body",
            "description": "dictionary containing the list of availability ids",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "availability_ids": {
                  "type": "list",
                  "example": ["aaaa11112223rf33451224", "438913477aaaaeff1498"]
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Availability deleted successfully"
          },
          "404": {
            "description": "Availability or Course not found"
          },
          "400": {
            "description": "Availability referenced by table(s)"
          }
        }
      },
      "post": {
        "summary": "Create a multiple(s) availabilities",
        "description": "Add availability or availabilities",
        "tags": ["Availability"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "description": "ID of the course",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "availability",
            "in": "body",
            "description": "Availability object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "availability_attr": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "start_time": {
                        "type": "string",
                        "example": "2024-03-11 13:07:56"
                      },
                      "day": {
                        "type": "string",
                        "example": "2024-03-11"
                      }
                    },
                    "required": ["start_time", "day"]
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Availability added successfully",
            "schema": {
              "$ref": "#/definitions/Availability"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Invalid course_id"
          }
        }
      }
      
    },
    "/api/v1/availability/{tutor_id}/tutor": {
      "get": {
        "summary": "Get all availabilities of a tutor",
        "description": "Retrieve all availabilities of a tutor.",
        "tags": ["Availability"],
        "parameters": [
          {
            "name": "tutor_id",
            "in": "path",
            "description": "ID of the tutor",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Availability"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tutor not found"
          }
        }
      }
    },      
...
  },
  "definitions": {
...
    "Availability": {
      "type": "object",
      "required": ["course_id", "start_time", "day"],
      "properties": {
        "course_id": {
          "type": "string",
          "example": "1234567890"
        },
        "start_time": {
          "type": "string",
          "format": "time",
          "example": "2024-02-29 12:30:00"
        },
        "end_time": {
          "type": "string",
          "format": "time",
          "example": "2024-02-29 2:30:00"
        },
        "day": {
          "type": "string",
          "format": "date",
          "example": "2024-02-29"
        }
      }
    }
  }
}